mv `scale_spectrum` to `surface.py`, rm unused functions
It was only used there, and the module is small enough that keeping it
elsewhere doesn't make much sense. Also removes one cross-import between
optics and effects. There is still `get_filter`, but maybe we can remove
that as well, and the whole scale function actually, if we used the
existing functionality in SpeXtra / synphot directly...

The functions `zero_mag_flux` and `get_zero_mag_spectrum` were not used
anywhere except in their own unit tests, and are definetely redundant with
SpeXtra / synphot, so were removed entirely.
