+++ scopesim/effects/ter_curves_utils.py
@@ -8,12 +8,8 @@ from astropy import units as u
 from astropy.table import Table
 from astropy.io.votable import parse_single_table
 from astropy.io import ascii as ioascii
-from astropy.io import fits
-from synphot import SpectralElement, SourceSpectrum, Empirical1D, Observation
-from synphot.units import PHOTLAM
+from synphot import SpectralElement, Empirical1D
 
-from scopesim.source.source_templates import vega_spectrum, st_spectrum, \
-    ab_spectrum
 from ..utils import find_file, quantity_from_table, get_logger
 from ..server.download_utils import create_client, handle_download
 
@@ -216,147 +212,6 @@ def get_filter(filter_name):
             filt = None
 
     return filt
-
-
-def get_zero_mag_spectrum(system_name: str = "AB") -> SourceSpectrum:
-    """
-    Return a synphot spectrum of 0 mag in given photometric system.
-
-    Parameters
-    ----------
-    system_name : {"AB", "Vega", "ST", "HST"}
-        Name of the photometric system.
-
-    Raises
-    ------
-    ValueError
-        Raised if `system_name` is invalid.
-
-    Returns
-    -------
-    spec : synphot.SourceSpectrum
-        Output spectrum.
-
-    """
-    if system_name.lower() == "vega":
-        return vega_spectrum()
-    if system_name.lower() == "ab":
-        return ab_spectrum()
-    if system_name.lower() in ("st", "hst"):
-        return st_spectrum()
-    raise ValueError(f"system_name {system_name} is unknown")
-
-
-def zero_mag_flux(filter_name, photometric_system, return_filter=False):
-    """
-    Return the zero magnitude photon flux for a filter.
-
-    Acceptable filter names are those given in
-    ``scopesim.effects.ter_curves_utils.FILTER_DEFAULTS`` or a string with an
-    appropriate name for a filter in the Spanish-VO filter-service. Such
-    strings must use the naming convention: observatory/instrument.filter. E.g:
-    "paranal/HAWKI.Ks", or "Gemini/GMOS-N.CaT".
-
-    Parameters
-    ----------
-    filter_name : str
-        Name of the filter - see above.
-
-    photometric_system : {"vega", "AB", "ST"}
-        Name of the photometric system
-
-    return_filter : bool, optional
-        If True, also returns the filter curve object
-
-    Returns
-    -------
-    flux : float
-        [PHOTLAM]
-    filt : ``synphot.SpectralElement``
-        If `return_filter` is True
-
-    """
-    filt = get_filter(filter_name)
-    spec = get_zero_mag_spectrum(photometric_system)
-
-    obs = Observation(spec, filt)
-    flux = obs.effstim(flux_unit=PHOTLAM)
-
-    if return_filter:
-        return flux, filt
-    return flux
-
-
-def scale_spectrum(spectrum, filter_name, amplitude):
-    """
-    Scale a SourceSpectrum to a value in a filter.
-
-    Parameters
-    ----------
-    spectrum : synphot.SourceSpectrum
-
-    filter_name : str
-        Name of a filter from
-        - a local instrument package (available in ``rc.__search_path__``)
-        - a generic filter name (see ``ter_curves_utils.FILTER_DEFAULTS``)
-        - a spanish-vo filter service reference (e.g. "Paranal/HAWKI.Ks")
-
-    amplitude : astropy.Quantity, float
-        The value that the spectrum should have in the given filter. Acceptable
-        astropy quantities are:
-        - u.mag : Vega magnitudes
-        - u.ABmag : AB magnitudes
-        - u.STmag : HST magnitudes
-        - u.Jy : Jansky per filter bandpass
-        Additionally the "FLAM" and "FNU" units from ``synphot.units`` can
-        be used when passing the quantity for `amplitude`.
-
-    Returns
-    -------
-    spectrum : synphot.SourceSpectrum
-        Input spectrum scaled to the given amplitude in the given filter.
-
-    Examples
-    --------
-    ::
-
-        >>> from scopesim.source.source_templates import vega_spectrum
-        >>> from scopesim.effects.ter_curves_utils as ter_utils
-        >>>
-        >>> spec = vega_spectrum()
-        >>> vega_185 = ter_utils.scale_spectrum(spec, "Ks", -1.85 * u.mag)
-        >>> ab_0 = ter_utils.scale_spectrum(spec, "Ks", 0 * u.ABmag)
-        >>> jy_3630 = ter_utils.scale_spectrum(spec, "Ks", 3630 * u.Jy)
-
-    """
-    if isinstance(amplitude, u.Quantity):
-        if amplitude.unit.physical_type == "spectral flux density":
-            if amplitude.unit != u.ABmag:
-                amplitude = amplitude.to(u.ABmag)
-            ref_spec = ab_spectrum(amplitude.value)
-
-        elif amplitude.unit.physical_type == "spectral flux density wav":
-            if amplitude.unit != u.STmag:
-                amplitude = amplitude.to(u.STmag)
-            ref_spec = st_spectrum(amplitude.value)
-
-        elif amplitude.unit == u.mag:
-            ref_spec = vega_spectrum(amplitude.value)
-
-        else:
-            raise ValueError(f"Units of amplitude must be one of "
-                             f"[u.mag, u.ABmag, u.STmag, u.Jy]: {amplitude}")
-    else:
-        ref_spec = vega_spectrum(amplitude)
-
-    filt = get_filter(filter_name)
-    ref_flux = Observation(ref_spec, filt).effstim(flux_unit=PHOTLAM)
-
-    real_flux = Observation(spectrum, filt).effstim(flux_unit=PHOTLAM)
-    scale_factor = ref_flux / real_flux
-    spectrum *= scale_factor.value
-
-    return spectrum
 
 
 def add_edge_zeros(tbl, wave_colname):
